requirements.txt
requests>=2.31
influxdb-client>=1.40
python-dotenv>=1.0
